requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
asyncpg
passlib[bcrypt]
psycopg2[binary]
python-jose[cryptography]
pytest
pytest-asyncio